zwischen Update vom Lern vortschritt + LohnRechner 23

+++ Function/tab.cpp
@@ -6,19 +6,19 @@ double tab( int tab, int index ) {
 		
 		if ( tab == 1 || tab == 4 ) {
 
-			return 0.400 - ( ( index - 1 ) * 0.016 );
+			return 0.400 - ( (double)( index - 1 ) * 0.016 );
 
 		} else if ( tab == 2 ) {
 
-			return 3000 - ( ( index - 1 ) * 120 );
+			return 3000 - ( (double)( index - 1 ) * 120 );
 
 		} else if ( tab == 3 ) {
 
-			return 900 - ( ( index - 1 ) * 36 );
+			return 900 - ( (double)( index - 1 ) * 36 );
 
 		} else if ( tab == 5 ) {
 
-			return 1900 - ( ( index - 1 ) * 76 );
+			return 1900 - ( (double)( index - 1 ) * 76 );
 
 		}
 
@@ -30,19 +30,19 @@ double tab( int tab, int index ) {
 
 		if ( tab == 1 || tab == 4 ) {
 
-			return 0.160 - ( ( index - 16 ) * 0.008 );
+			return 0.160 - ( (double)( index - 16 ) * 0.008 );
 
 		} else if ( tab == 2 ) {
 
-			return 1200 - ( ( index - 16 ) * 60 );
+			return 1200 - ( (double)( index - 16 ) * 60 );
 
 		} else if ( tab == 3 ) {
 
-			return 360 - ( ( index - 16 ) * 18 );
+			return 360 - ( (double)( index - 16 ) * 18 );
 
 		} else if ( tab == 5 ) {
 
-			return 760 - ( ( index - 16 ) * 38 );
+			return 760 - ( (double)( index - 16 ) * 38 );
 
 		}
